Extract TestFields out of AddTestForm
