+++ src/testing/AddTestForm.tsx
@@ -1,9 +1,11 @@
 import React from 'react';
-import { Box, Text, Label, Input, Checkbox, Button, Textarea } from 'theme-ui';
+import { Box, Button } from 'theme-ui';
 import { useFormik } from 'formik';
 import { Message } from 'retranslate';
 
-import { CreateTestCommand, TestType, FilledSchema, ObjectSchema, FieldSchema } from '../api';
+import { CreateTestCommand, TestType, FilledSchema } from '../api';
+
+import { getInitialValues, TestFields } from './TestFields';
 
 const AnyBox = Box as any;
 
@@ -17,7 +19,7 @@ export const AddTestForm = ({
   testType: TestType;
   onComplete: (command: CreateTestCommand) => any;
 }) => {
-  const form = useFormik({
+  const form: any = useFormik({
     initialValues: {
       ...getInitialValues(testType.resultsSchema),
       notes: '',
@@ -30,42 +32,7 @@ export const AddTestForm = ({
   return (
     <>
       <AnyBox as="form" sx={{ display: 'grid', gridGap: 4 }} onSubmit={form.handleSubmit}>
-        {Object.entries(testType.resultsSchema.properties).map(([key, value]) => {
-          if (value.type === 'boolean') {
-            return (
-              <Box key={key}>
-                <Label>
-                  <Checkbox {...form.getFieldProps(key)} />
-                  <Box>
-                    {value.title}
-                    {value.description ? (
-                      <Text as="small" sx={{ display: 'block' }}>
-                        {value.description}
-                      </Text>
-                    ) : null}
-                  </Box>
-                </Label>
-              </Box>
-            );
-          }
-
-          return (
-            <Box key={key}>
-              <Label htmlFor={`test-${key}`}>{value.title}</Label>
-              <Input
-                type={value.type === 'number' ? 'number' : 'text'}
-                id={`test-${key}`}
-                {...form.getFieldProps(key)}
-              />
-            </Box>
-          );
-        })}
-        <Box>
-          <Label htmlFor="test-notes">
-            <Message>addTestForm.notes.label</Message>
-          </Label>
-          <Textarea id="test-notes" {...form.getFieldProps('notes')} sx={{ resize: 'vertical' }} />
-        </Box>
+        <TestFields form={form} testType={testType} />
 
         <Button type="submit" variant="block">
           <Message>addTestForm.button</Message>
@@ -74,25 +41,3 @@ export const AddTestForm = ({
     </>
   );
 };
-
-function getInitialValues(schema: ObjectSchema) {
-  const initialValueEntries = Object.entries(schema.properties).map(([key, value]) => [
-    key,
-    initialPropertyValue(value),
-  ]);
-  return Object.fromEntries(initialValueEntries);
-}
-
-function initialPropertyValue({ type }: FieldSchema) {
-  switch (type) {
-    case 'boolean':
-      return false;
-    case 'number':
-      return 0;
-    case 'null':
-      return null;
-    case 'string': // fallthrough
-    default:
-      return '';
-  }
-}
