+++ src/testing/index.ts
@@ -2,3 +2,4 @@ export * from './TestList';
 export * from './AddTestPage';
 export * from './TestDetailPage';
 export * from './AddTestToIdentifierPage';
+export * from './TestFields';
